FEAT: Define ORM classes corresponding to DB Tables
Create the variables for each ORM class.
One variable for each column in a table.
Each ORM class will correspond to a table in the database.
Create a Foreign Key variable (column) on one of the ORM classes.
In the class pointed to by the Foreign Key - create a 'relationship' variable.
In this relationship variable declare a backreference which points to this ORM class.
This make this ORM class the 'many' side of a many-to-one relationship.