FEAT: Edit an existing database record
This is the 'U' in CRUD.
Pass the id of the Category selected for editing to the Controller method.
Retrieve the details of this Category from the database.
Pass the ORM Category object to the 'Edit Category' page.
Edit Category page shows the select category name which can be edited.
In the Controller method when the form data is POSTed back, set the name of the
 category ORM object to that passed in from the form.
Perform 'session.db.commit()'.